Fold ReferenceSerializationService into SerializerService
