+++ src/Workspaces/CoreTest/Execution/SnapshotSerializationTests.cs
@@ -465,14 +465,14 @@ MefHostServices.DefaultAssemblies.Add(typeof(Host.TemporaryStorageServiceFactory
             using var workspace = new AdhocWorkspace(hostServices);
             var reference = CreateShadowCopiedAnalyzerReference(tempRoot);
 
-            var assetBuilder = new CustomAssetBuilder(workspace);
-            var asset = assetBuilder.Build(reference, CancellationToken.None);
+            var serializer = workspace.Services.GetService<ISerializerService>();
+
+            var asset = WorkspaceAnalyzerReferenceAsset.Create(reference, serializer, CancellationToken.None);
 
             // verify checksum from custom asset builder uses different checksum than regular one
-            var service = workspace.Services.GetService<IReferenceSerializationService>();
             var expectedChecksum = Checksum.Create(
                 WellKnownSynchronizationKind.AnalyzerReference,
-                service.CreateChecksum(reference, CancellationToken.None));
+                serializer.CreateChecksum(reference, CancellationToken.None));
             Assert.Equal(expectedChecksum, asset.Checksum);
         }
